prepare deprecation of simple and alternative
