+++ test/test_bracketing.jl
@@ -436,8 +436,4 @@ end
     x = find_zero(f, (xleft, xright))
     @test abs(f(x)) <= 2eps(BigFloat)
 
-    # simple a42()
-    m = run_tests(Roots.a42)
-    VERSION >= v"1.6" && @test isempty(m.failures)
-    @test m.evalcount <= 3000 # paper says 2884, this has 2877
 end
