+++ test/test_find_zero.jl
@@ -429,9 +429,6 @@ end
         @test @inferred(find_zero(x -> x + 1, (-1, 1))) == -1
     end
 
-    ## issue #178 passing through method
-    @test fzero(sin, 3, 4, Roots.Brent()) ≈ π
-
     ## issue #188 with A42
     f = let a = 0.18
         x -> x * (1 - x^2) / ((x^2 + a^2) * (1 + a^2 * x^2))
