+++ test/test_fzero.jl
@@ -67,4 +67,8 @@ import Roots.fzero
 
     ### issue with fzeros and roots near 'b'
     @test 0 < maximum(fzeros(x -> sin(x) - 1 / 1000 * x, 0, pi)) < pi
+
+    ## issue #178 passing through method
+    @test fzero(sin, 3, 4, Roots.Brent()) ≈ π
+
 end
